requirements.txt
Flask==3.0.0
Flask-CORS==6.0.1
Werkzeug==3.0.1
requests==2.31.0
PyJWT==2.8.0
bcrypt==4.0.1
python-dateutil==2.9.0.post0
